Fixed weird bug with functions in util file
